refactor(balance): move folding to Snapshots
Move new_snapshots/ec_set_snapshots from Balances to snapshot.rs as
Snapshots::from_entries and Snapshots::from_ec_entries. This removes
the Self:: smell and places snapshot construction with the other
helpers.

Extract the version-chaining logic into a SnapshotFold accumulator so
it lives in one place instead of being copy-pasted across both entry
points. The two functions now differ only in whether they fold the
leaf account, stated inline as the domain rule it is. Restore the
tracing spans dropped in the initial move.

Add unit tests for EC set snapshot folding:
- updating an existing EC set balance
- fanning one member into multiple EC ancestor sets
